+++ src/kit/core/App.cpp
@@ -83,9 +83,10 @@ void App::initVita2d() {
 void App::debug() {
 
     ui->texts->drawF(0, 0, Body1, (unsigned int)RGBA8(255, 0, 0, 255), false,
-            "Current view: *%s*\nView found: %s",
+            "Current view: *%s*\nView found: %s\n%s",
             viewsController->getActualView().c_str(),
-            this->views.find(viewsController->getActualView()) != this->views.end() ? "true" : "false"
+            this->views.find(viewsController->getActualView()) != this->views.end() ? "true" : "false",
+            this->utils->PTC->debug().c_str()
     );
     int i = 0;
     for(auto const &kv : views) {