+++ src/kit/utils/UtilsPadTouchCtrl.cpp
@@ -3,11 +3,13 @@
 UtilsPadTouchCtrl::UtilsPadTouchCtrl(UtilsPad *pad) {
     touchMode = false;
     ctrlMode = true;
+    mode = PADTOUCHCTRL_MODE_CTRL;
 }
 
 
 UtilsPadTouchCtrl::UtilsPadTouchCtrl(UtilsPad *pad, UtilsTouch *touch) : pad(pad), touch(touch) {
     model = sceKernelGetModel();
+    mode = PADTOUCHCTRL_MODE_AUTO;
 
     if (model == SCE_KERNEL_MODEL_VITATV) {
         touchMode = false;
@@ -22,7 +24,7 @@ UtilsPadTouchCtrl::UtilsPadTouchCtrl(UtilsPad *pad, UtilsTouch *touch) : pad(pad
 void UtilsPadTouchCtrl::controller() {
 
     if (this->touch != nullptr) {
-        if (this->touch->clicking && ctrlMode) {
+        if (this->touch->clicking && ctrlMode && mode == PADTOUCHCTRL_MODE_AUTO) {
             yOldItem = yItem;
             xOldItem = xItem;
 
@@ -34,7 +36,7 @@ void UtilsPadTouchCtrl::controller() {
         }
     }
 
-    if (this->pad->held.clicking && touchMode) {
+    if (this->pad->held.clicking && touchMode && mode == PADTOUCHCTRL_MODE_AUTO) {
         touchMode = false;
         ctrlMode = true;
 
@@ -57,18 +59,18 @@ void UtilsPadTouchCtrl::controller() {
             xItem--;
         }
 
-        if (yItem > yLimit) {
+        if (yItem > yGlobalLimit) {
             yItem = 1;
         }
         if (yItem < 1) {
-            yItem = yLimit;
+            yItem = yGlobalLimit;
         }
 
-        if (xItem > xLimit) {
+        if (xItem > xGlobalLimit) {
             xItem = 1;
         }
         if (xItem < 1) {
-            xItem = xLimit;
+            xItem = xGlobalLimit;
         }
 
         for (auto& limit: yLimits) {
@@ -122,10 +124,10 @@ void UtilsPadTouchCtrl::setLimit(int xLimit, int yLimit, int xStart, int yStart)
 
 void UtilsPadTouchCtrl::updateLimit(PadTouchCtrlType type, int limit) {
     if (type == PADTOUCHCTRL_TYPE_X) {
-        xLimit = limit;
+        xGlobalLimit = limit;
     }
     else {
-        yLimit = limit;
+        yGlobalLimit = limit;
     }
 }
 
@@ -145,8 +147,8 @@ void UtilsPadTouchCtrl::updateStart(PadTouchCtrlType type, int start) {
 }
 
 void UtilsPadTouchCtrl::updateLimit(int xLimit, int yLimit) {
-    this->xOldItem = xLimit;
-    this->yOldItem = yLimit;
+    this->xGlobalLimit = xLimit;
+    this->yGlobalLimit = yLimit;
 }
 
 void UtilsPadTouchCtrl::updateStart(int xStart, int yStart) {
@@ -174,14 +176,14 @@ bool UtilsPadTouchCtrl::isX(PadTouchCtrlIs x) {
     if (x == PADTOUCHCTRL_IS_FIRST) {
         return xItem == 1;
     }
-    return xItem == xLimit;
+    return xItem == xGlobalLimit;
 }
 
 bool UtilsPadTouchCtrl::isY(PadTouchCtrlIs y) {
     if (y == PADTOUCHCTRL_IS_FIRST) {
         return yItem == 1;
     }
-    return yItem == yLimit;
+    return yItem == yGlobalLimit;
 }
 
 bool UtilsPadTouchCtrl::isXY(PadTouchCtrlIs x, PadTouchCtrlIs y) {
@@ -206,6 +208,40 @@ void UtilsPadTouchCtrl::addLimit(PadTouchCtrlType type, int line, int first, int
     }
 }
 
+void UtilsPadTouchCtrl::clearLimits() {
+    xLimits.clear();
+    yLimits.clear();
+}
+
+std::string UtilsPadTouchCtrl::debug() {
+    debugText = "x,y=";
+    debugText += std::to_string(xItem) + "," + std::to_string(yItem);
+
+    debugText = "x,y old=";
+    debugText += std::to_string(xOldItem) + "," + std::to_string(yOldItem);
+
+    debugText += "\nGlobalLimit x,y=";
+    debugText += std::to_string(xGlobalLimit) + "," + std::to_string(yGlobalLimit);
+
+    debugText += "\nxLimits: ";
+    for (auto& limit: xLimits) {
+        debugText += std::to_string(limit.first) + "=";
+        debugText += std::to_string(limit.second.first) + "," + std::to_string(limit.second.second) + "|";
+    }
+
+    debugText += "\nyLimits: ";
+    for (auto& limit: yLimits) {
+        debugText += std::to_string(limit.first) + "=";
+        debugText += std::to_string(limit.second.first) + "," + std::to_string(limit.second.second) + "|";
+    }
+
+    debugText += "\nMode: ";
+    debugText += this->isCtrlMode() ? "ctrl" : "touch";
+    debugText += this->getMode() == PADTOUCHCTRL_MODE_AUTO ? "|auto" : this->getMode() == PADTOUCHCTRL_MODE_CTRL ? "|ctrl" : "|touch";
+
+
+    return  debugText;
+}
 
 bool UtilsPadTouchCtrl::isTouchMode() const {
     return touchMode;
@@ -215,10 +251,20 @@ bool UtilsPadTouchCtrl::isCtrlMode() const {
     return ctrlMode;
 }
 
-void UtilsPadTouchCtrl::clearLimits() {
-    xLimits.clear();
-    yLimits.clear();
+PadTouchCtrlMode UtilsPadTouchCtrl::getMode() {
+    return mode;
 }
 
+void UtilsPadTouchCtrl::setMode(PadTouchCtrlMode mode) {
+    this->mode = mode;
 
+    if (this->mode == PADTOUCHCTRL_MODE_CTRL) {
+        ctrlMode = true;
+        touchMode = false;
+    }
+    else if (this->mode == PADTOUCHCTRL_MODE_TOUCH) {
+        touchMode = true;
+        ctrlMode = false;
+    }
+}