+++ src/kit/utils/UtilsPadTouchCtrl.hh
@@ -2,6 +2,7 @@
 #define VITAMATERIALKIT_PADTOUCHCTRL_HH
 
 #include <psp2/kernel/sysmem.h>
+#include <string>
 #include <list>
 
 #include "UtilsPad.hpp"
@@ -17,14 +18,22 @@ typedef enum PadTouchCtrlIs {
     PADTOUCHCTRL_IS_LAST
 } PadTouchCtrlIs;
 
+typedef enum PadTouchCtrlMode {
+    PADTOUCHCTRL_MODE_AUTO,
+    PADTOUCHCTRL_MODE_TOUCH,
+    PADTOUCHCTRL_MODE_CTRL
+} PadTouchCtrlMode;
+
 class UtilsPadTouchCtrl {
 private:
     UtilsPad *pad;
     UtilsTouch *touch;
     int model;
-    int xItem, yItem, xLimit, yLimit;
+    int xItem, yItem, xGlobalLimit, yGlobalLimit;
     int xOldItem, yOldItem;
     bool touchMode, ctrlMode;
+    PadTouchCtrlMode mode;
+    std::string debugText;
 
     std::list <std::pair<int, std::pair<int, int> >> xLimits, yLimits;
     void clearLimits();
@@ -58,9 +67,13 @@ public:
     bool isXY(int x, PadTouchCtrlIs y);
     bool isXY(PadTouchCtrlIs x, int y);
 
-
     bool isTouchMode() const;
     bool isCtrlMode() const;
+
+    PadTouchCtrlMode getMode();
+    void setMode(PadTouchCtrlMode mode);
+
+    std::string debug();
 };